requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
colorlog==6.7.0
exceptiongroup==1.1.1
h11==0.14.0
humanfriendly==10.0
idna==3.4
outcome==1.2.0
packaging==23.1
pip==23.0.1
pycparser==2.21
pyreadline3==3.4.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.28.2
selenium==4.9.0
setuptools==66.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
webdriver-manager==3.8.6
wheel==0.38.4
wsproto==1.2.0
